SIG: Move go/sig/metrics to go/sig/internal/metrics
