+++ go/sig/egress/dispatcher/BUILD.bazel
@@ -11,7 +11,7 @@ go_library(
         "//go/lib/log:go_default_library",
         "//go/lib/ringbuf:go_default_library",
         "//go/sig/egress/iface:go_default_library",
-        "//go/sig/metrics:go_default_library",
+        "//go/sig/internal/metrics:go_default_library",
         "//go/sig/mgmt:go_default_library",
     ],
 )
