+++ go/sig/egress/reader/BUILD.bazel
@@ -11,6 +11,6 @@ go_library(
         "//go/lib/ringbuf:go_default_library",
         "//go/sig/egress/iface:go_default_library",
         "//go/sig/egress/router:go_default_library",
-        "//go/sig/metrics:go_default_library",
+        "//go/sig/internal/metrics:go_default_library",
     ],
 )
