+++ go/sig/egress/session/BUILD.bazel
@@ -24,7 +24,7 @@ go_library(
         "//go/sig/egress/siginfo:go_default_library",
         "//go/sig/egress/worker:go_default_library",
         "//go/sig/internal/disp:go_default_library",
-        "//go/sig/metrics:go_default_library",
+        "//go/sig/internal/metrics:go_default_library",
         "//go/sig/mgmt:go_default_library",
         "//go/sig/sigcmn:go_default_library",
     ],
