+++ go/sig/egress/worker/BUILD.bazel
@@ -17,7 +17,7 @@ go_library(
         "//go/lib/util:go_default_library",
         "//go/sig/egress/iface:go_default_library",
         "//go/sig/egress/siginfo:go_default_library",
-        "//go/sig/metrics:go_default_library",
+        "//go/sig/internal/metrics:go_default_library",
         "//go/sig/mgmt:go_default_library",
         "//go/sig/sigcmn:go_default_library",
     ],
