+++ go/sig/ingress/BUILD.bazel
@@ -20,7 +20,7 @@ go_library(
         "//go/lib/snet:go_default_library",
         "//go/lib/sock/reliable:go_default_library",
         "//go/lib/util:go_default_library",
-        "//go/sig/metrics:go_default_library",
+        "//go/sig/internal/metrics:go_default_library",
         "//go/sig/mgmt:go_default_library",
         "//go/sig/sigcmn:go_default_library",
     ],
