+++ go/sig/internal/metrics/metrics.go
